src/game_service/requirements.txt
grpcio
grpcio-tools
flask
flask_cors